requirements.txt
asgiref==3.8.1
Django==4.2.20
django-admin-interface==3.0.0
django-colorfield==0.8.0
mysql==0.0.3
mysqlclient==2.2.7
pillow==11.2.1
python-slugify==8.0.4
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.13.2
tzdata==2025.2
